Add interface assertion and move recaptcha implementation to separate file
